pc-server/requirements.txt
sounddevice>=0.4.6
numpy>=1.24.0
opuslib>=3.0.1
pynput>=1.7.6
pystray>=0.19.5
Pillow>=10.0.0
